+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented in this file.
 - Description: Go1
 - Description: Laikago MJCF
 - Description: Panda MJCF
+- Description: Pepper
 - Description: Robotiq 2F-85
 - Description: Robotiq 2F-85 MJCF
 - Description: Shadow Hand MJCF